Johannes corrections
